+++ prisma/migrations/20260908081633_add_admin_totp_fields/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "Admin" ADD COLUMN     "backupCodes" TEXT[] DEFAULT ARRAY[]::TEXT[],
+ADD COLUMN     "totpEnabled" BOOLEAN NOT NULL DEFAULT false,
+ADD COLUMN     "totpLastUsedStep" INTEGER,
+ADD COLUMN     "totpSecret" TEXT;